+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/GeneratedCameraXLibrary.java
@@ -76,6 +76,23 @@ public class GeneratedCameraXLibrary {
     }
   }
 
+  /**
+   * The types (T) properly wrapped to be used as a LiveData<T>.
+   *
+   * <p>If you need to add another type to support a type S to use a LiveData<S> in this plugin,
+   * ensure the following is done on the Dart side:
+   *
+   * <p>* In `../lib/src/live_data.dart`, add new cases for S in
+   * `_LiveDataHostApiImpl#getValueFromInstances` to get the current value of type S from a
+   * LiveData<S> instance and in `LiveDataFlutterApiImpl#create` to create the expected type of
+   * LiveData<S> when requested.
+   *
+   * <p>On the native side, ensure the following is done:
+   *
+   * <p>* Update `LiveDataHostApiImpl#getValue` is updated to properly return identifiers for
+   * instances of type S. * Update `ObserverFlutterApiWrapper#onChanged` to properly handle
+   * receiving calls with instances of type S if a LiveData<S> instance is observed.
+   */
   public enum LiveDataSupportedType {
     CAMERA_STATE(0),
     ZOOM_STATE(1);
@@ -1297,8 +1314,6 @@ public class GeneratedCameraXLibrary {
       switch (type) {
         case (byte) 128:
           return ResolutionInfo.fromList((ArrayList<Object>) readValue(buffer));
-        case (byte) 129:
-          return ResolutionInfo.fromList((ArrayList<Object>) readValue(buffer));
         default:
           return super.readValueOfType(type, buffer);
       }
@@ -1308,9 +1323,6 @@ public class GeneratedCameraXLibrary {
     protected void writeValue(@NonNull ByteArrayOutputStream stream, Object value) {
       if (value instanceof ResolutionInfo) {
         stream.write(128);
-        writeValue(stream, ((ResolutionInfo) value).toList());
-      } else if (value instanceof ResolutionInfo) {
-        stream.write(129);
         writeValue(stream, ((ResolutionInfo) value).toList());
       } else {
         super.writeValue(stream, value);
@@ -1322,9 +1334,7 @@ public class GeneratedCameraXLibrary {
   public interface PreviewHostApi {
 
     void create(
-        @NonNull Long identifier,
-        @Nullable Long rotation,
-        @Nullable ResolutionInfo targetResolution);
+        @NonNull Long identifier, @Nullable Long rotation, @Nullable Long resolutionSelectorId);
 
     @NonNull
     Long setSurfaceProvider(@NonNull Long identifier);
@@ -1351,12 +1361,14 @@ public class GeneratedCameraXLibrary {
                 ArrayList<Object> args = (ArrayList<Object>) message;
                 Number identifierArg = (Number) args.get(0);
                 Number rotationArg = (Number) args.get(1);
-                ResolutionInfo targetResolutionArg = (ResolutionInfo) args.get(2);
+                Number resolutionSelectorIdArg = (Number) args.get(2);
                 try {
                   api.create(
                       (identifierArg == null) ? null : identifierArg.longValue(),
                       (rotationArg == null) ? null : rotationArg.longValue(),
-                      targetResolutionArg);
+                      (resolutionSelectorIdArg == null)
+                          ? null
+                          : resolutionSelectorIdArg.longValue());
                   wrapped.add(0, null);
                 } catch (Throwable exception) {
                   ArrayList<Object> wrappedError = wrapError(exception);
@@ -1911,40 +1923,11 @@ public class GeneratedCameraXLibrary {
           channelReply -> callback.reply(null));
     }
   }
-
-  private static class ImageCaptureHostApiCodec extends StandardMessageCodec {
-    public static final ImageCaptureHostApiCodec INSTANCE = new ImageCaptureHostApiCodec();
-
-    private ImageCaptureHostApiCodec() {}
-
-    @Override
-    protected Object readValueOfType(byte type, @NonNull ByteBuffer buffer) {
-      switch (type) {
-        case (byte) 128:
-          return ResolutionInfo.fromList((ArrayList<Object>) readValue(buffer));
-        default:
-          return super.readValueOfType(type, buffer);
-      }
-    }
-
-    @Override
-    protected void writeValue(@NonNull ByteArrayOutputStream stream, Object value) {
-      if (value instanceof ResolutionInfo) {
-        stream.write(128);
-        writeValue(stream, ((ResolutionInfo) value).toList());
-      } else {
-        super.writeValue(stream, value);
-      }
-    }
-  }
-
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface ImageCaptureHostApi {
 
     void create(
-        @NonNull Long identifier,
-        @Nullable Long flashMode,
-        @Nullable ResolutionInfo targetResolution);
+        @NonNull Long identifier, @Nullable Long flashMode, @Nullable Long resolutionSelectorId);
 
     void setFlashMode(@NonNull Long identifier, @NonNull Long flashMode);
 
@@ -1952,7 +1935,7 @@ public class GeneratedCameraXLibrary {
 
     /** The codec used by ImageCaptureHostApi. */
     static @NonNull MessageCodec<Object> getCodec() {
-      return ImageCaptureHostApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
     /**
      * Sets up an instance of `ImageCaptureHostApi` to handle messages through the
@@ -1970,12 +1953,14 @@ public class GeneratedCameraXLibrary {
                 ArrayList<Object> args = (ArrayList<Object>) message;
                 Number identifierArg = (Number) args.get(0);
                 Number flashModeArg = (Number) args.get(1);
-                ResolutionInfo targetResolutionArg = (ResolutionInfo) args.get(2);
+                Number resolutionSelectorIdArg = (Number) args.get(2);
                 try {
                   api.create(
                       (identifierArg == null) ? null : identifierArg.longValue(),
                       (flashModeArg == null) ? null : flashModeArg.longValue(),
-                      targetResolutionArg);
+                      (resolutionSelectorIdArg == null)
+                          ? null
+                          : resolutionSelectorIdArg.longValue());
                   wrapped.add(0, null);
                 } catch (Throwable exception) {
                   ArrayList<Object> wrappedError = wrapError(exception);
@@ -2038,6 +2023,182 @@ public class GeneratedCameraXLibrary {
 
                 api.takePicture(
                     (identifierArg == null) ? null : identifierArg.longValue(), resultCallback);
+              });
+        } else {
+          channel.setMessageHandler(null);
+        }
+      }
+    }
+  }
+
+  private static class ResolutionStrategyHostApiCodec extends StandardMessageCodec {
+    public static final ResolutionStrategyHostApiCodec INSTANCE =
+        new ResolutionStrategyHostApiCodec();
+
+    private ResolutionStrategyHostApiCodec() {}
+
+    @Override
+    protected Object readValueOfType(byte type, @NonNull ByteBuffer buffer) {
+      switch (type) {
+        case (byte) 128:
+          return ResolutionInfo.fromList((ArrayList<Object>) readValue(buffer));
+        default:
+          return super.readValueOfType(type, buffer);
+      }
+    }
+
+    @Override
+    protected void writeValue(@NonNull ByteArrayOutputStream stream, Object value) {
+      if (value instanceof ResolutionInfo) {
+        stream.write(128);
+        writeValue(stream, ((ResolutionInfo) value).toList());
+      } else {
+        super.writeValue(stream, value);
+      }
+    }
+  }
+
+  /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
+  public interface ResolutionStrategyHostApi {
+
+    void create(
+        @NonNull Long identifier, @Nullable ResolutionInfo boundSize, @Nullable Long fallbackRule);
+
+    /** The codec used by ResolutionStrategyHostApi. */
+    static @NonNull MessageCodec<Object> getCodec() {
+      return ResolutionStrategyHostApiCodec.INSTANCE;
+    }
+    /**
+     * Sets up an instance of `ResolutionStrategyHostApi` to handle messages through the
+     * `binaryMessenger`.
+     */
+    static void setup(
+        @NonNull BinaryMessenger binaryMessenger, @Nullable ResolutionStrategyHostApi api) {
+      {
+        BasicMessageChannel<Object> channel =
+            new BasicMessageChannel<>(
+                binaryMessenger, "dev.flutter.pigeon.ResolutionStrategyHostApi.create", getCodec());
+        if (api != null) {
+          channel.setMessageHandler(
+              (message, reply) -> {
+                ArrayList<Object> wrapped = new ArrayList<Object>();
+                ArrayList<Object> args = (ArrayList<Object>) message;
+                Number identifierArg = (Number) args.get(0);
+                ResolutionInfo boundSizeArg = (ResolutionInfo) args.get(1);
+                Number fallbackRuleArg = (Number) args.get(2);
+                try {
+                  api.create(
+                      (identifierArg == null) ? null : identifierArg.longValue(),
+                      boundSizeArg,
+                      (fallbackRuleArg == null) ? null : fallbackRuleArg.longValue());
+                  wrapped.add(0, null);
+                } catch (Throwable exception) {
+                  ArrayList<Object> wrappedError = wrapError(exception);
+                  wrapped = wrappedError;
+                }
+                reply.reply(wrapped);
+              });
+        } else {
+          channel.setMessageHandler(null);
+        }
+      }
+    }
+  }
+  /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
+  public interface ResolutionSelectorHostApi {
+
+    void create(
+        @NonNull Long identifier,
+        @Nullable Long resolutionStrategyIdentifier,
+        @Nullable Long aspectRatioStrategyIdentifier);
+
+    /** The codec used by ResolutionSelectorHostApi. */
+    static @NonNull MessageCodec<Object> getCodec() {
+      return new StandardMessageCodec();
+    }
+    /**
+     * Sets up an instance of `ResolutionSelectorHostApi` to handle messages through the
+     * `binaryMessenger`.
+     */
+    static void setup(
+        @NonNull BinaryMessenger binaryMessenger, @Nullable ResolutionSelectorHostApi api) {
+      {
+        BasicMessageChannel<Object> channel =
+            new BasicMessageChannel<>(
+                binaryMessenger, "dev.flutter.pigeon.ResolutionSelectorHostApi.create", getCodec());
+        if (api != null) {
+          channel.setMessageHandler(
+              (message, reply) -> {
+                ArrayList<Object> wrapped = new ArrayList<Object>();
+                ArrayList<Object> args = (ArrayList<Object>) message;
+                Number identifierArg = (Number) args.get(0);
+                Number resolutionStrategyIdentifierArg = (Number) args.get(1);
+                Number aspectRatioStrategyIdentifierArg = (Number) args.get(2);
+                try {
+                  api.create(
+                      (identifierArg == null) ? null : identifierArg.longValue(),
+                      (resolutionStrategyIdentifierArg == null)
+                          ? null
+                          : resolutionStrategyIdentifierArg.longValue(),
+                      (aspectRatioStrategyIdentifierArg == null)
+                          ? null
+                          : aspectRatioStrategyIdentifierArg.longValue());
+                  wrapped.add(0, null);
+                } catch (Throwable exception) {
+                  ArrayList<Object> wrappedError = wrapError(exception);
+                  wrapped = wrappedError;
+                }
+                reply.reply(wrapped);
+              });
+        } else {
+          channel.setMessageHandler(null);
+        }
+      }
+    }
+  }
+  /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
+  public interface AspectRatioStrategyHostApi {
+
+    void create(
+        @NonNull Long identifier, @NonNull Long preferredAspectRatio, @NonNull Long fallbackRule);
+
+    /** The codec used by AspectRatioStrategyHostApi. */
+    static @NonNull MessageCodec<Object> getCodec() {
+      return new StandardMessageCodec();
+    }
+    /**
+     * Sets up an instance of `AspectRatioStrategyHostApi` to handle messages through the
+     * `binaryMessenger`.
+     */
+    static void setup(
+        @NonNull BinaryMessenger binaryMessenger, @Nullable AspectRatioStrategyHostApi api) {
+      {
+        BasicMessageChannel<Object> channel =
+            new BasicMessageChannel<>(
+                binaryMessenger,
+                "dev.flutter.pigeon.AspectRatioStrategyHostApi.create",
+                getCodec());
+        if (api != null) {
+          channel.setMessageHandler(
+              (message, reply) -> {
+                ArrayList<Object> wrapped = new ArrayList<Object>();
+                ArrayList<Object> args = (ArrayList<Object>) message;
+                Number identifierArg = (Number) args.get(0);
+                Number preferredAspectRatioArg = (Number) args.get(1);
+                Number fallbackRuleArg = (Number) args.get(2);
+                try {
+                  api.create(
+                      (identifierArg == null) ? null : identifierArg.longValue(),
+                      (preferredAspectRatioArg == null)
+                          ? null
+                          : preferredAspectRatioArg.longValue(),
+                      (fallbackRuleArg == null) ? null : fallbackRuleArg.longValue());
+                  wrapped.add(0, null);
+                } catch (Throwable exception) {
+                  ArrayList<Object> wrappedError = wrapError(exception);
+                  wrapped = wrappedError;
+                }
+                reply.reply(wrapped);
               });
         } else {
           channel.setMessageHandler(null);
@@ -2194,37 +2355,10 @@ public class GeneratedCameraXLibrary {
           channelReply -> callback.reply(null));
     }
   }
-
-  private static class ImageAnalysisHostApiCodec extends StandardMessageCodec {
-    public static final ImageAnalysisHostApiCodec INSTANCE = new ImageAnalysisHostApiCodec();
-
-    private ImageAnalysisHostApiCodec() {}
-
-    @Override
-    protected Object readValueOfType(byte type, @NonNull ByteBuffer buffer) {
-      switch (type) {
-        case (byte) 128:
-          return ResolutionInfo.fromList((ArrayList<Object>) readValue(buffer));
-        default:
-          return super.readValueOfType(type, buffer);
-      }
-    }
-
-    @Override
-    protected void writeValue(@NonNull ByteArrayOutputStream stream, Object value) {
-      if (value instanceof ResolutionInfo) {
-        stream.write(128);
-        writeValue(stream, ((ResolutionInfo) value).toList());
-      } else {
-        super.writeValue(stream, value);
-      }
-    }
-  }
-
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface ImageAnalysisHostApi {
 
-    void create(@NonNull Long identifier, @Nullable ResolutionInfo targetResolutionIdentifier);
+    void create(@NonNull Long identifier, @Nullable Long resolutionSelectorId);
 
     void setAnalyzer(@NonNull Long identifier, @NonNull Long analyzerIdentifier);
 
@@ -2232,7 +2366,7 @@ public class GeneratedCameraXLibrary {
 
     /** The codec used by ImageAnalysisHostApi. */
     static @NonNull MessageCodec<Object> getCodec() {
-      return ImageAnalysisHostApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
     /**
      * Sets up an instance of `ImageAnalysisHostApi` to handle messages through the
@@ -2250,11 +2384,13 @@ public class GeneratedCameraXLibrary {
                 ArrayList<Object> wrapped = new ArrayList<Object>();
                 ArrayList<Object> args = (ArrayList<Object>) message;
                 Number identifierArg = (Number) args.get(0);
-                ResolutionInfo targetResolutionIdentifierArg = (ResolutionInfo) args.get(1);
+                Number resolutionSelectorIdArg = (Number) args.get(1);
                 try {
                   api.create(
                       (identifierArg == null) ? null : identifierArg.longValue(),
-                      targetResolutionIdentifierArg);
+                      (resolutionSelectorIdArg == null)
+                          ? null
+                          : resolutionSelectorIdArg.longValue());
                   wrapped.add(0, null);
                 } catch (Throwable exception) {
                   ArrayList<Object> wrappedError = wrapError(exception);